+++ Changelog.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 ### Fixed
 ### Removed
+### Deprecated
+- The `convert-periods-to-pocket` command will be removed in version `1.0.0`.
 
 ## [v0.26.2.1] - 2021-12-16
 ### Fixed