+++ Changelog.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support updating `end` field for recurrent entries as unset (i.e. indicating no end).
 - Support updating `category` field for entries as unset (i.e. indicating unspecified category).
 ### Changed
+- Use `black` as code formatter.
 ### Fixed
 ### Removed
 - Remove support for Python 3.5 for good (broken since v0.26.1.0 anyways). (#59)